+++ apps/platform/api-shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @doccov/api-shared
 
+## 0.2.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @doccov/sdk@0.30.0
+
 ## 0.2.25
 
 ### Patch Changes